Fix file paths reported in ls
Incorrect file paths were being reported in `hadoop fs -ls`. Fixed this.
For this, had to change the file path formats for files within a tar
archive. For example, say there is a file with path `dir1/file1` inside
a tar archive `archive.tar`. Following are the paths in old format and
new format:

  - Old format: `tar:///archive.tar+dir1/file1`
  - New format: `tar:///archive.tar/+dir1+file1`

This new format is not supported throughout. Both in CLI and APIs.