Fix issue in sub-directory check in TarIndex
The TarIndex component of TarFileSystem does not yet support
directories. So while creating index, the directory entries are skipped.
The check for directory previously was `entry.isFile()` but that does
not work in all cases.

It was observed that for the directory entries both
`entry.isDirectory()` and `entry.isFile()` was true. So, changed the
directory check in TarIndex to check for both isFile and isDirectory.

This should fix the `StackOverflowError` seen in `issues#6`